+++ tests/integration/Espo/Core/Utils/FieldManager/EnumTypeTest.php
@@ -51,9 +51,21 @@ class EnumTypeTest extends \tests\integration\Core\BaseTestCase
         "tooltip":false
     }';
 
+    protected function createFieldManager($app = null)
+    {
+        if (!$app) {
+            $app = $this;
+        }
+
+        return $app->getContainer()->get('injectableFactory')->create(
+            'Espo\\Tools\\FieldManager\\FieldManager'
+        );
+    }
+
     public function testCreate()
     {
-        $fieldManager = $this->getContainer()->get('fieldManager');
+        $fieldManager = $this->createFieldManager();
+
         $fieldDefs = get_object_vars(json_decode($this->jsonFieldDefs));
 
         $fieldManager->create('Account', 'testEnum', $fieldDefs);
@@ -71,6 +83,7 @@ class EnumTypeTest extends \tests\integration\Core\BaseTestCase
         $this->assertTrue($savedFieldDefs['audited']);
 
         $entityManager = $app->getContainer()->get('entityManager');
+
         $account = $entityManager->getEntity('Account');
         $account->set([
             'name' => 'Test',
@@ -88,7 +101,9 @@ class EnumTypeTest extends \tests\integration\Core\BaseTestCase
         $this->testCreate();
 
         $app = $this->createApplication();
-        $fieldManager = $app->getContainer()->get('fieldManager');
+
+        $fieldManager = $this->createFieldManager($app);
+
         $fieldDefs = get_object_vars(json_decode($this->jsonFieldDefs));
         $fieldDefs['required'] = false;
         $fieldDefs['default'] = 'option3';
