+++ tests/integration/Espo/Core/Utils/FieldManager/VarcharTypeTest.php
@@ -49,9 +49,21 @@ class VarcharTypeTest extends \tests\integration\Core\BaseTestCase
         "tooltip":false
     }';
 
+    protected function createFieldManager($app = null)
+    {
+        if (!$app) {
+            $app = $this;
+        }
+
+        return $app->getContainer()->get('injectableFactory')->create(
+            'Espo\\Tools\\FieldManager\\FieldManager'
+        );
+    }
+
     public function testCreate()
     {
-        $fieldManager = $this->getContainer()->get('fieldManager');
+        $fieldManager = $this->createFieldManager();
+
         $fieldDefs = get_object_vars(json_decode($this->jsonFieldDefs));
 
         $fieldManager->create('Account', 'testVarchar', $fieldDefs);
@@ -86,7 +98,9 @@ class VarcharTypeTest extends \tests\integration\Core\BaseTestCase
         $this->testCreate();
 
         $app = $this->createApplication();
-        $fieldManager = $app->getContainer()->get('fieldManager');
+
+        $fieldManager = $this->createFieldManager($app);
+
         $fieldDefs = get_object_vars(json_decode($this->jsonFieldDefs));
         $fieldDefs['required'] = false;
         $fieldDefs['default'] = 'default-value';
@@ -103,6 +117,7 @@ class VarcharTypeTest extends \tests\integration\Core\BaseTestCase
         $this->assertEquals('default-value', $savedFieldDefs['default']);
 
         $entityManager = $app->getContainer()->get('entityManager');
+
         $account = $entityManager->getEntity('Account');
         $account->set([
             'name' => 'New Test',
